% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight6.R
\docType{data}
\name{flights6}
\alias{flights6}
\title{Flights data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2020656 rows and 19 columns.}
\usage{
flights6
}
\description{
On-time data for all flights that departed NYC (i.e. JFK, LGA or EWR) in
2013.
}
\keyword{datasets}
