% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_month_data}
\alias{get_month_data}
\title{Month data}
\usage{
get_month_data(month)
}
\arguments{
\item{month}{Scalar integer defining the month.}
}
\value{
A \code{data.frame} with the same structure as \code{\link[nycflights13:flights]{nycflights13::flights}}.
}
\description{
Returns flight data for the given \code{month}.
}
\examples{
summary(get_month_data(6))

}
