% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_report}
\alias{render_report}
\title{Render Report}
\usage{
render_report(report, ..., output_dir = ".")
}
\arguments{
\item{report}{The name of the report, without the \code{".Rmd"} extension.}

\item{...}{Additional arguments to be passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}

\item{output_dir}{The output directory for the rendered report.}
}
\value{
The path to the produced report.
}
\description{
Render a given report to HTML.
}
\examples{
output <- render_report(
  "ith-report", params = list(i = 9), output_dir = tempfile()
)
if (interactive()) browseURL(output)

}
