% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_path}
\alias{report_path}
\title{Report Path}
\usage{
report_path(report)
}
\arguments{
\item{report}{The name of the report, without the \code{".Rmd"} extension.}
}
\value{
The path to the desired R Markdown file.
}
\description{
Construct the path to the R Markdown source file of a given report.
}
